requirements.txt
gql==3.5.2
pydantic==2.11.3
python-dotenv==1.1.0
Requests==2.32.3
urllib3==2.4.0
